Implement CCTV addon
This is an addon that introduces two new items:
1. CCTV Camera
2. CCTV Base Station (Laptop)

Players can drop the CCTV Cameras throughout the map and use the CCTV Base
Station to remote monitor the cameras.

While dropping the cameras, the player can choose who is allowed to access
the camera:

1. Player (himself)
2. Group (his group)
3. Team (his team)

Config allows admin to set the maximum CCTV range.

Conflicts:
	addons/R3F_ARTY_AND_LOG/R3F_LOG/config.sqf
	client/init.sqf
	client/items/init.sqf
	storeConfig.sqf