Add saving for CCTV cameras
This persistence is specific to the Node.js (sock-rpc-stats)
method. If you are trying to add persistence for iniDB, or extDB
you need to make sure that the followin variables are saving with the
camera object ( and restored as well).

a3w_cctv_camera
camera_name
camera_owner_type
camera_owner_value

Conflicts:
	server/default_config.sqf