Remove satchels, claymores, and demo charges
Even though we can detect these, and save them to the database there is no
easy way to re-create them.

I added function there for re-creating them based on a KillZoneKid's
article. However, it's super hacky ... so I am reluctant to do it that
way. Maybe in the future, BIS will make an easier way for creating those.

Conflicts:
	server/functions/serverVars.sqf